requirements.txt
accelerate==0.30.1
aiohttp==3.9.3
aiosignal==1.3.1
alabaster==0.7.16
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
autopep8==2.0.4
babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
calflops==0.2.9
captum==0.2.0
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
cilog==1.3.0
click==8.1.7
cloudpickle==3.0.0
comm==0.2.1
conda-pack==0.8.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
# dgl==2.2.0+cu121 -f https://data.dgl.ai/wheels/torch-2.1/cu121/repo.html
dglgo==0.0.2
docker-pycreds==0.4.0
docutils==0.20.1
easydict==1.13
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fast-pagerank==1.0.0
fastjsonschema==2.19.1
filelock==3.9.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.5.0
gdown==5.1.0
geomloss==0.2.6
gitdb==4.0.11
gitpython==3.1.41
h11==0.14.0
h5py==3.10.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.23.1
hydra-core==1.3.2
idna==3.4
imagesize==1.4.1
ipdb==0.13.13
ipykernel==6.29.2
ipython==8.21.0
ipywidgets==8.1.2
isoduration==20.11.0
isort==5.13.2
jedi==0.19.1
jinja2==3.1.2
joblib==1.3.2
json5==0.9.17
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-client==8.6.0
jupyter-console==6.6.3
jupyter-core==5.7.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter-server==2.12.5
jupyter-server-terminals==0.5.2
jupyterlab==4.1.2
jupyterlab-pygments==0.3.0
jupyterlab-server==2.25.3
jupyterlab-widgets==3.0.10
lightning-utilities==0.10.1
littleutils==0.2.2
llvmlite==0.42.0
markupsafe==2.1.3
matplotlib==3.9.0
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.5
munch==4.0.0
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
notebook==7.1.0
notebook-shim==0.2.4
numba==0.59.1
numpy==1.26.3
numpydoc==1.6.0
ogb==1.3.6
omegaconf==2.3.0
openpyxl==3.1.2
outdated==0.2.2
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandoc==2.3
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.2.0
plumbum==1.8.2
ply==3.11
prometheus-client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.6.1
pydantic-core==2.16.2
pyemd==1.0.0
# pyg-lib==0.4.0+pt21cu121 -f https://data.pyg.org/whl/torch-2.1.0+cu121.html
pygments==2.17.2
pygmtools==0.5.0
pyparsing==3.1.1
pyscf==2.5.0
pysocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2024.1
pyyaml==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
qtpy==2.4.1
rdkit-pypi==2022.9.5
referencing==0.33.0
regex==2024.5.15
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
ruamel-yaml==0.18.6
ruamel-yaml-clib==0.2.8
safetensors==0.4.3
scikit-learn==1.4.0
scikit-network==0.32.1
scipy==1.12.0
seaborn==0.13.2
send2trash==1.8.2
sentry-sdk==1.40.4
setproctitle==1.3.3
shap==0.45.0
six==1.16.0
slicer==0.0.7
smmap==5.0.1
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.5
sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tensorboardx==2.6.2.2
terminado==0.18.0
threadpoolctl==3.3.0
tinycss2==1.2.1
tokenizers==0.19.1
tomli==2.0.1
# torch-cluster==1.6.3+pt21cu121 -f https://data.pyg.org/whl/torch-2.1.0+cu121.html
torch-geometric==2.5.2
# torch-scatter==2.1.2+pt21cu121 -f https://data.pyg.org/whl/torch-2.1.0+cu121.html
# torch-sparse==0.6.18+pt21cu121 -f https://data.pyg.org/whl/torch-2.1.0+cu121.html
# torch-spline-conv==1.2.2+pt21cu121 -f https://data.pyg.org/whl/torch-2.1.0+cu121.html
# torchaudio==2.1.2+cu121
torchdata==0.7.1
torchmetrics==1.3.1
# torchvision==0.16.2+cu121
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.41.0
triton==2.1.0
typed-argument-parser==1.7.2
typer==0.9.0
types-python-dateutil==2.8.19.20240106
typing-extensions==4.8.0
typing-inspect==0.9.0
tzdata==2024.1
ucimlrepo==0.0.6
uri-template==1.3.0
urllib3==1.26.13
wandb==0.16.3
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
widgetsnbextension==4.0.10
yarl==1.9.4